+++ Core/Diagrams/Path.cs
@@ -107,7 +107,7 @@ namespace Cession.Diagrams
             return false;
         }
 
-        internal override void DoOffset (int x, int y)
+        internal override void DoOffset (double x, double y)
         {
             _segments.ForEach (s => s.DoOffset (x, y));
         }
